+++ test/index.js
@@ -4,33 +4,42 @@ const path = require('path'),
   chai = require('chai'),
   should = chai.should(),
   chaiAsPromised = require('chai-as-promised'),
-  SERVERLESS_PATH = path.join(process.cwd(), 'node_modules', 'serverless', 'lib'),
-  Serverless = require('serverless'),
-  utils = require(path.join(SERVERLESS_PATH, 'utils'));
+  Serverless = require('serverless');
 
 chai.use(chaiAsPromised);
 
-const JSHintPlugin = require('..')(
-  require(path.join(SERVERLESS_PATH, 'ServerlessPlugin')),
-  SERVERLESS_PATH
-);
-
-let s, plugin, logs;
+let s, plugin, logs, JSHintPlugin;
 
 const logger = function(log) {
   logs.push(log);
 };
 
 describe('ServerlessJSHint', function() {
-  before(function() {
+  beforeEach(function(done) {
     this.timeout(0);
-    s = new Serverless();
-    plugin = new JSHintPlugin(s, { logger: logger });
-    s.addPlugin(plugin);
-  });
 
-  beforeEach(function() {
+    s = new Serverless();
     logs = [];
+
+    s.init().then(function() {
+      JSHintPlugin = require('..')(s);
+      plugin = new JSHintPlugin({ logger: logger });
+
+      s.addPlugin(plugin);
+      s.config.projectPath = path.join(__dirname, 'test-prj');
+      s.setProject(new s.classes.Project({
+        stages: {
+          dev: { regions: { 'eu-west-1': {} }}
+        },
+        variables: {
+          project: 'serverless-project',
+          stage: 'dev',
+          region: 'eu-west-1'
+        }
+      }));
+
+      done();
+    });
   });
 
   describe('#getName()', function() {
@@ -48,49 +57,51 @@ describe('ServerlessJSHint', function() {
 
   describe('#functionJSHint()', function() {
     it('should fail for non-existing functions', function() {
-      return plugin.functionJSHint({ options: { path: 'nodejs/someFunction' }}).should.be.rejected.then(function() {
+      return plugin.functionJSHint({ options: { names: ['someFunction'] }}).should.be.rejected.then(function() {
         logs.should.be.empty;
       });
     });
 
     it('should succeed for valid functions', function() {
-      s.config.projectPath = path.join(__dirname, 'test-prj');
-      s.state.setAsset(new s.classes.Component(s, { sPath: 'nodejs' }));
-      s.state.setAsset(new s.classes.Function(s, { sPath: 'nodejs/validFunction' }));
+      _bootstrapFunction('validNodeFunction', 'nodejs');
 
-      return plugin.functionJSHint({ options: { path: 'nodejs/validFunction' }}).should.be.fulfilled.then(function() {
+      return plugin.functionJSHint({ options: { names: ['validNodeFunction'] }}).should.be.fulfilled.then(function() {
         logs[0].should.contain('Success!');
       });
     });
 
     it('should report errors for invalid functions', function() {
-      s.config.projectPath = path.join(__dirname, 'test-prj');
-      s.state.setAsset(new s.classes.Component(s, { sPath: 'nodejs' }));
-      s.state.setAsset(new s.classes.Function(s, { sPath: 'nodejs/invalidFunction' }));
+      _bootstrapFunction('invalidNodeFunction', 'nodejs');
 
-      return plugin.functionJSHint({ options: { path: 'nodejs/invalidFunction' }}).should.be.fulfilled.then(function() {
+      return plugin.functionJSHint({ options: { names: ['invalidNodeFunction'] }}).should.be.fulfilled.then(function() {
         logs[0].should.contain('Error!');
       });
     });
 
     it('should apply a custom configuration file', function() {
       s.config.projectPath = path.join(__dirname, 'test-prj-2');
-      s.state.setAsset(new s.classes.Component(s, { sPath: 'nodejs' }));
-      s.state.setAsset(new s.classes.Function(s, { sPath: 'nodejs/curlyFunction' }));
+      _bootstrapFunction('curlyFunction', 'nodejs');
 
-      return plugin.functionJSHint({ options: { path: 'nodejs/curlyFunction' }}).should.be.fulfilled.then(function() {
+      return plugin.functionJSHint({ options: { names: ['curlyFunction'] }}).should.be.fulfilled.then(function() {
         logs[0].should.contain('Error!');
       });
     });
 
     it('should fail on non-Node components', function() {
-      let component = new s.classes.Component(s, { sPath: 'python' });
-      component.runtime = 'python2.7';
+      _bootstrapFunction('pythonFunction', 'python2.7');
 
-      s.state.setAsset(component);
-      s.state.setAsset(new s.classes.Function(s, { sPath: 'python/pyFunction' }));
-
-      return plugin.functionJSHint({ options: { path: 'python/pyFunction' }}).should.be.rejected;
+      return plugin.functionJSHint({ options: { names: ['pythonFunction'] }}).should.be.rejected;
     });
   });
 });
+
+function _bootstrapFunction(name, runtime) {
+  const func = new s.classes.Function({
+    name: name,
+    runtime: runtime
+  }, path.join(s.config.projectPath, name, 's-function.json'));
+
+  s.getProject().setFunction(func);
+
+  return func;
+}