+++ test/test-prj/validNodeFunction/handler.js
@@ -3,6 +3,6 @@
 module.exports.handler = function(event, context) {
   var response = {};
   if (response) response.foo = 'bar';
-  
+
   context.done(null, response);
 };