+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.0
+
+* Made compatible with Serverless v0.5
+* Accept multiple names as arguments
+
 ## 0.0.2
 
 * Support for custom configuration file (.jshintrc)